Abstract duplicate for maven and npm and use original data format for the declared_license of npm  #1403
