+++ src/packagedcode/npm.py
@@ -41,6 +41,7 @@ from six import string_types
 from commoncode import filetype
 from commoncode import fileutils
 from packagedcode import models
+from packagedcode.utils import compute_normalized_license
 from packagedcode.utils import parse_repo_url
 
 
@@ -97,64 +98,6 @@ class NpmPackage(models.Package):
 
     def compute_normalized_license(self):
         return compute_normalized_license(self.declared_license)
-
-def compute_normalized_license(declared_license):
-    """
-    Return a detected license expression from a declared license mapping.
-    """
-    if not declared_license:
-        return
-
-    licensing = Licensing()
-
-    detected_licenses = []
-
-    for license_declaration in declared_license:
-        # 1. try detection on the value of type if not empty and keep this
-        type = license_declaration.get('type')
-        via_type = models.compute_normalized_license(type)
-
-        # 2. try detection on the value of url  if not empty and keep this
-        url = license_declaration.get('url')
-        via_url = models.compute_normalized_license(url)
-
-
-        if via_type:
-            # The type should have precedence and any unknowns
-            # in url should be ignored.
-            if via_url == 'unknown':
-                via_url = None
-
-        # Check the three detections to decide which license to keep
-        type_and_url = via_type == via_url
-        all_same = type_and_url
-        if via_type:
-            if all_same:
-                detected_licenses.append(via_type)
-            else:
-                # we have some non-unknown license detected in url
-                detections = via_type, via_url
-                detections = [l for l in detections if l]
-                if detections:
-                    if len(detections) == 1:
-                        combined_expression = detections[0]
-                    else:
-                        expressions = [
-                            licensing.parse(le, simple=True) for le in detections]
-                        combined_expression = str(licensing.AND(*expressions))
-                    detected_licenses.append(combined_expression)
-
-        elif via_url:
-            detected_licenses.append(via_url)
-
-    if len(detected_licenses) == 1:
-        return detected_licenses[0]
-
-    if detected_licenses:
-        # Combine if pom contains more than one licenses declarations.
-        expressions = [licensing.parse(le, simple=True) for le in detected_licenses]
-        combined_expression = licensing.AND(*expressions)
-        return str(combined_expression)
 
 
 def npm_homepage_url(namespace, name, registry='https://www.npmjs.com/package'):
@@ -411,7 +354,7 @@ def get_declared_licenses(license_object):
         # TODO: handle "SEE LICENSE IN <filename>"
         # TODO: handle UNLICENSED
         # TODO: parse expression with license_expression library
-        declared_licenses.append({'type': license_object})
+        return license_object
 
     elif isinstance(license_object, dict):
         # old, deprecated form
@@ -431,15 +374,7 @@ def get_declared_licenses(license_object):
         or
         "licenses": ["MIT"],
         """
-        for lic in license_object:
-            if not lic:
-                continue
-            if isinstance(lic, string_types):
-                declared_licenses.append({'type': lic})
-            elif isinstance(lic, dict):
-                declared_licenses.append(lic)
-            else:
-                declared_licenses.append({'type': repr(lic)})
+        declared_licenses.extend(license_object)
     return declared_licenses
 
 
@@ -459,9 +394,8 @@ def licenses_mapper(license, licenses, package):  # NOQA
     -  "license": "UNLICENSED" means this is proprietary
     """
     declared_license = get_declared_licenses(license)
-    licenses = get_declared_licenses(licenses)
-    declared_license.extend(licenses)
-
+    if not declared_license:
+        declared_license = get_declared_licenses(licenses)
     package.declared_license = declared_license
     return package
