+++ src/packagedcode/utils.py
@@ -28,6 +28,10 @@ from __future__ import unicode_literals
 
 from six import string_types
 
+from license_expression import Licensing
+
+from packagedcode import models
+
 
 PLAIN_URLS = (
     'https://',
@@ -146,3 +150,91 @@ def build_description(summary, description):
             description = '\n'.join([summary , description])
 
     return description
+
+
+def compute_normalized_license(declared_license):
+    """
+    Return a detected license expression from a declared license mapping.
+    This is for maven and npm packages.
+    """
+    if not declared_license:
+        return
+
+    licensing = Licensing()
+
+    detected_licenses = []
+    
+    if isinstance(declared_license, string_types):
+        # if the declared_license is a string, use list to wrapper it and use the following loop to handle in order not to have to many if condition
+        declared_license = [declared_license]
+    for license_declaration in declared_license:
+        if isinstance(license_declaration, string_types):
+            license_declaration = models.compute_normalized_license(license_declaration)
+            if license_declaration:
+                detected_licenses.append(license_declaration)
+        elif isinstance(license_declaration, dict):
+            # 1. try detection on the value of name if not empty and keep this
+            name = license_declaration.get('name')
+            if not name:
+                # this is for npm, npm has type field instead of name
+                name = license_declaration.get('type')
+            via_name = models.compute_normalized_license(name)
+    
+            # 2. try detection on the value of url  if not empty and keep this
+            url = license_declaration.get('url')
+            via_url = models.compute_normalized_license(url)
+    
+            # 3. try detection on the value of comment  if not empty and keep this
+            comments = license_declaration.get('comments')
+            via_comments = models.compute_normalized_license(comments)
+    
+    
+            if via_name:
+                # The name should have precedence and any unknowns
+                # in url and comment should be ignored.
+                if via_url == 'unknown':
+                    via_url = None
+                if via_comments == 'unknown':
+                    via_comments = None
+    
+            # Check the three detections to decide which license to keep
+            name_and_url = via_name == via_url
+            name_and_comment = via_name == via_comments
+            all_same = name_and_url and name_and_comment
+    
+            if via_name:
+                if all_same:
+                    detected_licenses.append(via_name)
+    
+                # name and (url or comment) are same
+                elif name_and_url and not via_comments:
+                    detected_licenses.append(via_name)
+                elif name_and_comment and not via_url:
+                    detected_licenses.append(via_name)
+    
+                else:
+                    # we have some non-unknown license detected in url or comment
+                    detections = via_name, via_url, via_comments
+                    detections = [l for l in detections if l]
+                    if detections:
+                        if len(detections) == 1:
+                            combined_expression = detections[0]
+                        else:
+                            expressions = [
+                                licensing.parse(le, simple=True) for le in detections]
+                            combined_expression = str(licensing.AND(*expressions))
+                        detected_licenses.append(combined_expression)
+    
+            elif via_url:
+                detected_licenses.append(via_url)
+            elif via_comments:
+                detected_licenses.append(via_comments)
+
+    if len(detected_licenses) == 1:
+        return detected_licenses[0]
+
+    if detected_licenses:
+        # Combine if pom contains more than one licenses declarations.
+        expressions = [licensing.parse(le, simple=True) for le in detected_licenses]
+        combined_expression = licensing.AND(*expressions)
+        return str(combined_expression)
