+++ tests/packagedcode/data/npm/as_installed/package.json.expected
@@ -2205,11 +2205,7 @@
   "vcs_url": "git+https://github.com/npm/npm.git@fc7bbf03e39cc48a8924b90696d28345a6a90f3c",
   "copyright": null,
   "license_expression": null,
-  "declared_license": [
-    {
-      "type": "Artistic-2.0"
-    }
-  ],
+  "declared_license": "Artistic-2.0",
   "notice_text": null,
   "manifest_path": null,
   "dependencies": [
