+++ tests/packagedcode/data/npm/invalid-dep/package.json.expected
@@ -39,11 +39,7 @@
   "vcs_url": "git+https://github.com/stevepapa/angular2-autosize.git",
   "copyright": null,
   "license_expression": null,
-  "declared_license": [
-    {
-      "type": "MIT"
-    }
-  ],
+  "declared_license": "MIT",
   "notice_text": null,
   "manifest_path": null,
   "dependencies": [
