+++ tests/packagedcode/data/npm/keywords/package.json.expected
@@ -54,11 +54,7 @@
   "vcs_url": "https://github.com/lodash/lodash",
   "copyright": null,
   "license_expression": null,
-  "declared_license": [
-    {
-      "type": "MIT"
-    }
-  ],
+  "declared_license": "MIT",
   "notice_text": null,
   "manifest_path": null,
   "dependencies": [
