+++ tests/packagedcode/data/npm/scoped1/package.json.expected
@@ -30,11 +30,7 @@
   "vcs_url": "git+https://github.com/ionic-team/ionic-app-scripts.git",
   "copyright": null,
   "license_expression": null,
-  "declared_license": [
-    {
-      "type": "MIT"
-    }
-  ],
+  "declared_license": "MIT",
   "notice_text": null,
   "manifest_path": null,
   "dependencies": [
