+++ tests/packagedcode/data/npm/scoped2/package.json.expected
@@ -30,11 +30,7 @@
   "vcs_url": "https://github.com/ivx/eslint-config-ivx.git",
   "copyright": null,
   "license_expression": null,
-  "declared_license": [
-    {
-      "type": "MIT"
-    }
-  ],
+  "declared_license": "MIT",
   "notice_text": null,
   "manifest_path": null,
   "dependencies": [
