+++ tests/packagedcode/data/npm/sequelize/package.json.expected
@@ -67,11 +67,7 @@
   "vcs_url": "git+https://github.com/sequelize/sequelize.git",
   "copyright": null,
   "license_expression": null,
-  "declared_license": [
-    {
-      "type": "MIT"
-    }
-  ],
+  "declared_license": "MIT",
   "notice_text": null,
   "manifest_path": null,
   "dependencies": [
